+++ src/main/java/org/jiranibora/com/contributions/ContributionService.java
@@ -16,7 +16,7 @@ import org.jiranibora.com.models.Member;
 import org.jiranibora.com.models.MemberContribution;
 import org.jiranibora.com.models.MemberContributionPK;
 import org.jiranibora.com.models.Penalty;
-
+import org.jiranibora.com.penalty.PenaltyRepository;
 import org.springframework.security.authentication.AnonymousAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
@@ -112,7 +112,7 @@ public class ContributionService {
 
     // Member's contribution
     @Transactional(rollbackOn = Exception.class)
-    public ContributionResponse memberContribution(String contributionId, Integer amount) throws Exception {
+    public ContributionResponse memberContribution(String contributionId, Double amount) throws Exception {
         ContributionResponse contRes = new ContributionResponse();
         // Is there any contribution with such ID
